requirements.txt
Django==2.0.3
django-solo==1.1.3
djangorestframework==3.7.7
